refactor(no-rule): change structure of the project
