+++ index.html
@@ -4,14 +4,14 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://fonts.googleapis.com/css?family=Lato:300,400,700" rel="stylesheet">
-    <link rel="stylesheet" href="style.css" media="screen">
+    <link rel="stylesheet" href="./styles/main.css" media="screen">
     <title>Todo App</title>
   </head>
   <body class="page">
     <header class="header wrapper">
       <img 
         class="header__img" 
-        src="./eisenhower-matrix.jpg" 
+        src="./assets/eisenhower-matrix.jpg" 
         alt="Eisenhower matrix diagram"
       >
       <a 
@@ -46,7 +46,7 @@
             >
               <img 
                 class="task__icon" 
-                src="./remove.svg" 
+                src="./assets/remove.svg" 
                 alt="Remove task button icon"
               >
             </button>
@@ -67,7 +67,7 @@
             >
               <img 
                 class="task__icon"
-                src="./remove.svg" 
+                src="./assets/remove.svg" 
                 alt="Remove task button icon"
               >
             </button>
@@ -90,7 +90,7 @@
             >
               <img 
                 class="task__icon" 
-                src="./remove.svg" 
+                src="./assets/remove.svg" 
                 alt="Remove task button icon"
               >
             </button>
@@ -98,6 +98,6 @@
         </ul>
       </section>
     </main>
-    <script type="module" src="app.js"></script>
+    <script type="module" src="./js/app.js"></script>
   </body>
 </html>
