--- style.css
@@ -1,187 +1,0 @@
-/*
---------------
-Reset
---------------
-*/
-*,
-::before,
-::after {
-  box-sizing: border-box;
-}
-
-body,
-h3, 
-p {
-  margin: 0;
-}
-
-ul {
-  margin: 0;
-  padding: 0px;
-  list-style: none;
-}
-
-input,
-button{
-  outline: none;
-}
-
-/*
---------------
-Base
---------------
-*/
-.page {
-  background-color: #f8f8f8;
-  color: #333;
-  font-family: Lato, sans-serif;
-}
-
-.wrapper {
-  width: 100vw;
-  min-width: 375px;
-  margin: 0 auto 0;
-  padding: 0 10px;
-}
-
-@media (min-width: 520px) {
-  .wrapper { 
-    width: 500px;
-    padding: 0;
-  }
-}
-
-.btn-primary {
-  background: none;
-  border: 0px;
-  color: #888;
-  font-size: 15px;
-  width: 60px;
-  font-family: Lato, sans-serif;
-  cursor: pointer;
-}
-
-.btn-primary:hover {
-  color: #3a3a3a;
-}
-
-.section-title {
-  color: #333;
-  font-weight: 700;
-  font-size: 15px;
-  border-bottom: 2px solid #333;
-  padding: 30px 0 10px;
-  margin: 0;
-  text-transform: uppercase;
-}
-
-.text-input-primary {
-  margin: 0;
-  font-size: 18px;
-  line-height: 18px;
-  height: 21px;
-  padding: 0 9px;
-  border: 1px solid #ddd;
-  background: #fff;
-  border-radius: 6px;
-  font-family: Lato, sans-serif;
-  color: #888;
-}
-
-.text-input-primary:focus {
-  color: #333;
-}
-
-/*
---------------
-Header
---------------
-*/
-.header {
-  text-align: right;
-}
-
-.header.wrapper {
-  padding-top: 10px;
-}
-
-.header__img {
-  width: 100%;
-}
-
-.header__more {
-  font-family: Georgia, 'Times New Roman', Times, serif;
-}
-
-@media (max-width: 768px) {
-  .header { 
-    text-align: center;
-  }
-}
-
-/*
---------------
-Task
---------------
-*/
-.task {
-  overflow: hidden;
-  padding: 20px 0;
-  border-bottom: 1px solid #eee;
-
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.task__check {
-  margin: 0 10px;
-}
-
-.task__input,
-.task__label {
-  width: 56%;
-  display: inline-block;
-  flex-grow: 1;
-}
-
-.task__icon {
-  height: 2em;
-  transform: rotateZ(45deg);
-  transition: transform 200ms ease-in;
-}
-
-.task__icon:hover {
-  transform: rotateZ(0);
-}
-
-.task_edit .task__input {
-  display: inline-block;
-  width: 224px;
-}
-
-.task__input,
-.task_edit .task__label {
-  display: none;
-}
-
-/*
---------------
-Misc
---------------
-*/
-.add-task__group {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  padding: 20px 0;
-}
-
-.add-task__input {
-  flex-grow: 1;
-}
-
-.completed__tasks .task__label {
-  text-decoration: line-through;
-  color: #888;
-}
