+++ styles/base.css
@@ -1,0 +1,60 @@
+.page {
+  background-color: #f8f8f8;
+  color: #333;
+  font-family: Lato, sans-serif;
+}
+
+.wrapper {
+  width: 100vw;
+  min-width: 375px;
+  margin: 0 auto 0;
+  padding: 0 10px;
+}
+
+@media (min-width: 520px) {
+  .wrapper { 
+    width: 500px;
+    padding: 0;
+  }
+}
+
+.section-title {
+  color: #333;
+  font-weight: 700;
+  font-size: 15px;
+  border-bottom: 2px solid #333;
+  padding: 30px 0 10px;
+  margin: 0;
+  text-transform: uppercase;
+}
+
+.btn-primary {
+  background: none;
+  border: 0px;
+  color: #888;
+  font-size: 15px;
+  width: 60px;
+  font-family: Lato, sans-serif;
+  cursor: pointer;
+}
+
+.btn-primary:hover {
+  color: #3a3a3a;
+}
+
+.text-input-primary {
+  margin: 0;
+  font-size: 18px;
+  line-height: 18px;
+  height: 21px;
+  padding: 0 9px;
+  border: 1px solid #ddd;
+  background: #fff;
+  border-radius: 6px;
+  font-family: Lato, sans-serif;
+  color: #888;
+}
+
+.text-input-primary:focus {
+  color: #333;
+}
