+++ styles/main.css
@@ -1,0 +1,4 @@
+@import url("./reset.css");
+@import url("./base.css");
+@import url("./header.css");
+@import url("./task.css");
