+++ styles/reset.css
@@ -1,0 +1,22 @@
+*,
+::before,
+::after {
+  box-sizing: border-box;
+}
+
+body,
+h3, 
+p {
+  margin: 0;
+}
+
+ul {
+  margin: 0;
+  padding: 0px;
+  list-style: none;
+}
+
+input,
+button{
+  outline: none;
+}
